+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 - 🩹 Shorten pipeline info in log for long config options
 - 🐛 Fix cached jobs being put into queue
 - 🩹 Shorten job debug messages when hit limits
+- 🚑 Remove sort_dicts for pprint.pformat for py3.7
 
 ## 0.2.13